src/requirements.txt
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.0
python-dotenv==1.1.1
scikit_learn==1.7.0
